feat: [split-factory-service-named-factory-editable-definition-cluster-001] - [Preserve named-factory creation and current-definition lookup while splitting ownership]
