+++ docs/internal/processes/development-guide-relevant-files.md
@@ -90,7 +90,7 @@ This inventory records the checked-in files and directories that the maintainer 
 | `pkg/generatedclient/` | Generated Go API client package | Backend functional tests that need typed client coverage should import this package instead of rebuilding `http.Get` or `http.Post` helpers for OpenAPI-owned routes. |
 | `pkg/factory/interfaces.go` | Shared API runtime contract owner | Shared API-facing runtime methods such as work submission, factory event streaming, and engine snapshot reads should stay owned here so stricter boundary seams can embed one canonical base contract instead of re-declaring duplicate method sets. |
 | `pkg/apisurface/contract.go` | API-server runtime seam owner | API-server-facing seams that add behavior on top of the shared runtime contract should embed `factory.APIFactory`, keep additive named-factory or explicit session-routing methods here, and avoid route-local duplicate interface ownership across packages. |
-| `pkg/service/factory.go` | Named-factory activation and current-runtime seam | `CreateNamedFactory`, `ActivateNamedFactory`, and `GetCurrentNamedFactory` own the persisted-factory, runtime-swap, and current-pointer contract; backend changes and tests should extend this service seam instead of duplicating activation rules in handlers or ad hoc helpers. When maintainability cleanup touches service startup or activation, keep default-runtime and replacement-runtime construction flowing through shared same-package helpers so logger setup, replay recorder wiring, worker loading, and inputs watcher behavior stay aligned across both paths. Service-owned long-lived sidecars such as file watchers, cron schedulers, and poller supervisors should also stay wired through `startLiveRuntimeSidecars(...)` and `stopLiveRuntimeSidecars(...)` here so lifecycle start, cancellation, replacement shutdown, and test harness setup continue to share one authoritative runtime-sidecar seam. Pollers that submit ingress work on behalf of one runtime, including hosted and script pollers, should receive the runtime-scoped submitter from this seam instead of rediscovering the current factory through global service state. |
+| `pkg/service/factory.go` and `pkg/service/factory_named_factory.go` | Named-factory activation and current-runtime seam | Keep persisted named-factory read and serialization ownership in `pkg/service/factory_named_factory.go`, while `pkg/service/factory.go` continues owning runtime activation, replacement-runtime startup, and long-lived sidecar lifecycle. Backend changes and tests should extend this same-package service seam instead of duplicating activation or current-pointer rules in handlers or ad hoc helpers. When maintainability cleanup touches service startup or activation, keep default-runtime and replacement-runtime construction flowing through shared same-package helpers so logger setup, replay recorder wiring, worker loading, and inputs watcher behavior stay aligned across both paths. Service-owned long-lived sidecars such as file watchers, cron schedulers, and poller supervisors should also stay wired through `startLiveRuntimeSidecars(...)` and `stopLiveRuntimeSidecars(...)` here so lifecycle start, cancellation, replacement shutdown, and test harness setup continue to share one authoritative runtime-sidecar seam. Pollers that submit ingress work on behalf of one runtime, including hosted and script pollers, should receive the runtime-scoped submitter from this seam instead of rediscovering the current factory through global service state. |
 | `pkg/service/runtime_sessions.go` | Live multi-session runtime lifecycle seam | Concurrent factory-session ownership, the reserved `~default` compatibility alias, folder-target discovery for new sessions, background runtime startup, and per-session stop behavior should stay centralized here so later session-scoped API, CLI, and UI work can reuse one live-session manager instead of rebuilding parallel runtime maps. Keep session-open startup on the split-context contract already shared with replacement runtime startup: the long-lived service context owns the background runtime and sidecars, but the original request context must still bound `waitForLiveRuntimeStart(...)` so canceled open-session requests do not leave behind newly registered live sessions. |
 | `pkg/service/factory_cron_scheduler_test.go` | Service-mode cron runtime harness owner | Service-mode cron ingress helpers should not return immediately after `go svc.Run(...)`; wait until `currentLiveRuntime()` exists and `waitForLiveRuntimeStart(...)` succeeds before driving submissions or shutdown, otherwise cancellation-oriented runtime fixes can surface as pre-start test races instead of real cron behavior regressions. |
 | `pkg/testutil/mock_factory.go` | Shared API/service seam test double | API and service seam tests should reuse this mock and keep its direct `factory.APIFactory`, `apisurface.APISurface`, and `apisurface.SessionAPISurface` compile-time assertions aligned with the live runtime contracts instead of building route-local stubs that can drift. |
